Set up inital state for dropdown
